backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.46
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
pandas==2.3.3
psycopg2-binary==2.9.11
gunicorn==23.0.0
python-dotenv==1.0.1
flask-cors==4.0.0
psycopg2-binary==2.9.11